requirements.txt
 pandas
 numpy
 seaborn
 scipy
 matplotlib-venn
 scikit-learn
 networkx
 pyvis
 pyarrow
